requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
PyPDF2
langchain
openai
faiss-cpu
langchain-community
tiktoken
langchain-openai
python-dotenv